% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings.R
\name{rankings}
\alias{rankings}
\title{Get ATP/WTA rankings}
\usage{
rankings(wta = NULL, decades = NULL)
}
\arguments{
\item{wta}{logical. Download women's rankings?}

\item{decades}{a vector of decade codes}
}
\value{
a \code{tibble} with data on historical ATP/WTA rankings
}
\description{
This function downloads ATP or WTA tennis rankings from \url{https://github.com/JeffSackmann}
}
\author{
Jeppe Vierø
}
